requirements.txt
debugpy==1.6.6
Django==4.1.5
django-bootstrap4==22.3
django-cors-headers==3.13.0
django-phonenumber-field==7.0.2
django_debug_toolbar==3.8.1
phonenumbers==8.13.7
phonenumberslite==8.13.7
Pillow==9.4.0
psycopg==3.1.8
psycopg2-binary==2.9.5
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
uri-template==1.2.0
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
Werkzeug==2.2.2
WTForms==3.0.1
gunicorn==20.1.0
whitenoise==6.4.0
sqlparse==0.2.4



